telnyx-twilio-migration/migration-examples/django/call-tracking-django/requirements.txt
# Requirements for the Call Tracking sample Django project

# Django (1.8 is a long-term support release)
Django==3.2.0

# Third-party app dependencies
django-bootstrap3==21.1.0
django-forms-bootstrap==3.1.0
django-phonenumber-field==0.7.2
phonenumbers==8.13.0
psycopg2==2.8.6
telnyx>=2.0,<3.0
whitenoise==5.3.0

# Production dependencies
dj-database-url==0.3.0
gunicorn==19.3.0

# Test dependencies
coverage==3.7.1
mock==1.3.0
model-mommy==2.0.0
six==1.9.0
